# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Single-user local Codex-like Agent API built with Go and CloudWeGo Eino.\nExternal effects always flow through ToolRouter, EffectInference, PolicyEngine, ApprovalCenter, and Executor.",
        "title": "Local Agent API",
        "contact": {},
        "version": "dev"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/runs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Runs"
                ],
                "summary": "List workflow runs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comma-separated run statuses",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max runs to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.RunListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/runs/{run_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Runs"
                ],
                "summary": "Get workflow run",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Run ID",
                        "name": "run_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/local-agent_internal_agent.RunState"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/runs/{run_id}/cancel": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Runs"
                ],
                "summary": "Cancel workflow run",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Run ID",
                        "name": "run_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/local-agent_internal_agent.RunState"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/runs/{run_id}/resume": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Runs"
                ],
                "summary": "Resume paused workflow run",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Run ID",
                        "name": "run_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Resume payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ResumeRunRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/local-agent_internal_agent.RunResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/runs/{run_id}/steps": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Runs"
                ],
                "summary": "List workflow run steps",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Run ID",
                        "name": "run_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.RunStepListResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_handlers.ResumeRunRequest": {
            "type": "object",
            "properties": {
                "approval_id": {
                    "type": "string"
                },
                "approved": {
                    "type": "boolean"
                }
            }
        },
        "internal_api_handlers.RunListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/local-agent_internal_agent.RunState"
                    }
                }
            }
        },
        "internal_api_handlers.RunStepListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/local-agent_internal_agent.RunStep"
                    }
                }
            }
        },
...
        "local-agent_internal_agent.AgentContext": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Message"
                    }
                }
            }
        },
        "local-agent_internal_agent.Plan": {
            "type": "object",
            "properties": {
                "decision": {
                    "$ref": "#/definitions/local-agent_internal_agent.PlanDecision"
                },
                "message": {
                    "type": "string"
                },
                "preamble": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "tool_proposal": {
                    "$ref": "#/definitions/local-agent_internal_core.ToolProposal"
                }
            }
        },
        "local-agent_internal_agent.PlanDecision": {
            "type": "string",
            "enum": [
                "answer",
                "tool",
                "continue",
                "stop"
            ],
            "x-enum-varnames": [
                "PlanDecisionAnswer",
                "PlanDecisionTool",
                "PlanDecisionContinue",
                "PlanDecisionStop"
            ]
        },
        "local-agent_internal_agent.RunResponse": {
            "type": "object",
            "properties": {
                "approval": {
                    "$ref": "#/definitions/local-agent_internal_core.ApprovalRecord"
                },
                "assistant_message": {
                    "$ref": "#/definitions/local-agent_internal_core.Message"
                },
                "run_id": {
                    "type": "string"
                },
                "state": {
                    "$ref": "#/definitions/local-agent_internal_agent.RunState"
                },
                "tool_result": {
                    "$ref": "#/definitions/local-agent_internal_core.ToolResult"
                }
            }
        },
        "local-agent_internal_agent.RunState": {
            "type": "object",
            "properties": {
                "approval_id": {
                    "type": "string"
                },
                "context": {
                    "$ref": "#/definitions/local-agent_internal_agent.AgentContext"
                },
                "conversation_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "current_step": {
                    "type": "string"
                },
                "current_step_index": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "inference": {
                    "$ref": "#/definitions/local-agent_internal_core.EffectInferenceResult"
                },
                "max_steps": {
                    "type": "integer"
                },
                "plan": {
                    "$ref": "#/definitions/local-agent_internal_agent.Plan"
                },
                "policy": {
                    "$ref": "#/definitions/local-agent_internal_core.PolicyDecision"
                },
                "proposal": {
                    "$ref": "#/definitions/local-agent_internal_core.ToolProposal"
                },
                "run_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/local-agent_internal_agent.RunStatus"
                },
                "step_count": {
                    "type": "integer"
                },
                "tool_result": {
                    "$ref": "#/definitions/local-agent_internal_core.ToolResult"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_message": {
                    "type": "string"
                }
            }
        },
        "local-agent_internal_agent.RunStatus": {
            "type": "string",
            "enum": [
                "received_user_message",
                "context_built",
                "planned",
                "tool_proposed",
                "effect_inferred",
                "policy_decided",
                "approval_requested",
                "paused_for_approval",
                "approval_approved",
                "approval_rejected",
                "tool_executing",
                "tool_completed",
                "tool_failed",
                "assistant_summarizing",
                "completed",
                "failed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "RunStatusReceived",
                "RunStatusContextBuilt",
                "RunStatusPlanned",
                "RunStatusToolProposed",
                "RunStatusEffectInferred",
                "RunStatusPolicyDecided",
                "RunStatusApprovalRequested",
                "RunStatusPausedForApproval",
                "RunStatusApprovalApproved",
                "RunStatusApprovalRejected",
                "RunStatusToolExecuting",
                "RunStatusToolCompleted",
                "RunStatusToolFailed",
                "RunStatusAssistantSummarize",
                "RunStatusCompleted",
                "RunStatusFailed",
                "RunStatusCancelled"
            ]
        },
        "local-agent_internal_agent.RunStep": {
            "type": "object",
            "properties": {
                "approval": {
                    "$ref": "#/definitions/local-agent_internal_core.ApprovalRecord"
                },
                "created_at": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "inference": {
                    "$ref": "#/definitions/local-agent_internal_core.EffectInferenceResult"
                },
                "policy": {
                    "$ref": "#/definitions/local-agent_internal_core.PolicyDecision"
                },
                "proposal": {
                    "$ref": "#/definitions/local-agent_internal_core.ToolProposal"
                },
                "run_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/local-agent_internal_agent.RunStepStatus"
                },
                "step_id": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "tool_result": {
                    "$ref": "#/definitions/local-agent_internal_core.ToolResult"
                },
                "type": {
                    "$ref": "#/definitions/local-agent_internal_agent.RunStepType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "local-agent_internal_agent.RunStepStatus": {
            "type": "string",
            "enum": [
                "pending",
                "completed",
                "paused",
                "failed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "RunStepStatusPending",
                "RunStepStatusCompleted",
                "RunStepStatusPaused",
                "RunStepStatusFailed",
                "RunStepStatusCancelled"
            ]
        },
        "local-agent_internal_agent.RunStepType": {
            "type": "string",
            "enum": [
                "build_context",
                "plan",
                "continue",
                "propose_tool",
                "infer_effect",
                "decide_policy",
                "request_approval",
                "execute_tool",
                "summarize",
                "finalize",
                "cancel"
            ],
            "x-enum-varnames": [
                "RunStepTypeBuildContext",
                "RunStepTypePlan",
                "RunStepTypeContinue",
                "RunStepTypeProposeTool",
                "RunStepTypeInferEffect",
                "RunStepTypeDecidePolicy",
                "RunStepTypeRequestApproval",
                "RunStepTypeExecuteTool",
                "RunStepTypeSummarize",
                "RunStepTypeFinalize",
                "RunStepTypeCancel"
            ]
        },
        "local-agent_internal_core.ApprovalRecord": {
            "type": "object",
            "properties": {
                "conversation_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "decision": {
                    "$ref": "#/definitions/local-agent_internal_core.PolicyDecision"
                },
                "id": {
                    "type": "string"
                },
                "inference": {
                    "$ref": "#/definitions/local-agent_internal_core.EffectInferenceResult"
                },
                "input_snapshot": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "proposal": {
                    "$ref": "#/definitions/local-agent_internal_core.ToolProposal"
                },
                "reason": {
                    "type": "string"
                },
                "resolved_at": {
                    "type": "string"
                },
                "run_id": {
                    "type": "string"
                },
                "snapshot_hash": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/local-agent_internal_core.ApprovalStatus"
                },
                "summary": {
                    "type": "string"
                }
            }
        },
        "local-agent_internal_core.ApprovalStatus": {
            "type": "string",
            "enum": [
                "pending",
                "approved",
                "rejected"
            ],
            "x-enum-varnames": [
                "ApprovalPending",
                "ApprovalApproved",
                "ApprovalRejected"
            ]
        },
...
        "local-agent_internal_core.EffectInferenceResult": {
            "type": "object",
            "properties": {
                "approval_required": {
                    "type": "boolean"
                },
                "confidence": {
                    "type": "number"
                },
                "effects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "reason_summary": {
                    "type": "string"
                },
                "risk_level": {
                    "type": "string"
                },
                "sensitive": {
                    "type": "boolean"
                }
            }
        },
...
        "local-agent_internal_core.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "content_json": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "conversation_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "local-agent_internal_core.PolicyDecision": {
            "type": "object",
            "properties": {
                "allowed": {
                    "type": "boolean"
                },
                "approval_payload": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "reason": {
                    "type": "string"
                },
                "requires_approval": {
                    "type": "boolean"
                },
                "risk_level": {
                    "type": "string"
                }
            }
        },
...
        "local-agent_internal_core.ToolProposal": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "expected_effects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "input": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "purpose": {
                    "type": "string"
                },
                "tool": {
                    "type": "string"
                }
            }
        },
        "local-agent_internal_core.ToolResult": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "finished_at": {
                    "type": "string"
                },
                "output": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "started_at": {
                    "type": "string"
                },
                "tool_call_id": {
                    "type": "string"
                }
            }
        },
...
        "schema.Message": {
            "type": "object",
            "properties": {
                "assistant_output_multi_content": {
                    "description": "AssistantGenMultiContent is for receiving multimodal output from the model.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.MessageOutputPart"
                    }
                },
                "content": {
                    "description": "Content is for user text input and model text output.",
                    "type": "string"
                },
                "extra": {
                    "description": "customized information for model implementation",
                    "type": "object",
                    "additionalProperties": {}
                },
                "multi_content": {
                    "description": "if MultiContent is not empty, use this instead of Content\nif MultiContent is empty, use Content\nDeprecated: Use UserInputMultiContent for user multimodal inputs and AssistantGenMultiContent for model multimodal outputs.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.ChatMessagePart"
                    }
                },
                "name": {
                    "type": "string"
                },
                "reasoning_content": {
                    "description": "ReasoningContent is the thinking process of the model, which will be included when the model returns reasoning content.",
                    "type": "string"
                },
                "response_meta": {
                    "$ref": "#/definitions/schema.ResponseMeta"
                },
                "role": {
                    "$ref": "#/definitions/schema.RoleType"
                },
                "tool_call_id": {
                    "description": "only for ToolMessage",
                    "type": "string"
                },
                "tool_calls": {
                    "description": "only for AssistantMessage",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.ToolCall"
                    }
                },
                "tool_name": {
                    "description": "only for ToolMessage",
                    "type": "string"
                },
                "user_input_multi_content": {
                    "description": "UserInputMultiContent passes multimodal content provided by the user to the model.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.MessageInputPart"
                    }
                }
            }
        },
...
        "schema.RoleType": {
            "type": "string",
            "enum": [
                "assistant",
                "user",
                "system",
                "tool"
            ],
            "x-enum-varnames": [
                "Assistant",
                "User",
                "System",
                "Tool"
            ]
        },
...
    }
}